feat: added parser and global variables
Yes, even tho a parser existed before I made small changes. It still uses toml++ under the hood, but I added error stuff and auto path. The only reason why this exists is that I now compressed 8 LoC into 1. The global variables is only a single one and that's for the dotfiles path, it eliminates long stuff and make it more centralized so it's easier to change later on. It's also prefixed to /home/calishu, so you can't just randomly put it at /boot