feat(config): added 'get_dependency' for information about it
This basically gives you the full path of the source and destination of a config. It takes up the dependency name and returns a unordered map with the source and destination. I also fixed some stuff in the 'get_dependencies', but I remembered it too late, so I need to include it in this commit