refactor(config): fixed a lot of stuff
e.g. removing the default argument on the definiton while the headers also had it defined. A different change is the structure of the class, previous I thought you could use that 'class [...]{}' could also use in the definition, but I was wrong and it caused a lot of errors.